% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_income_df_long.R
\docType{data}
\name{average_income_df_long}
\alias{average_income_df_long}
\title{Average Income dataframe 1}
\format{
A tibble
\describe{
}
}
\source{
\url{}
}
\usage{
average_income_df_long
}
\description{
Average Income dataframe
}
\keyword{datasets}
