% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime_df.R
\docType{data}
\name{crime_df}
\alias{crime_df}
\title{Crime dataframe}
\format{
A tibble
\describe{
}
}
\source{
\url{https://www.pnas.org/content/114/7/1512/tab-figures-data}
}
\usage{
crime_df
}
\description{
Crime dataframe
}
\keyword{datasets}
