% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrimeIncomeDf-package.R
\docType{package}
\name{CrimeIncomeDf-package}
\alias{CrimeIncomeDf}
\alias{CrimeIncomeDf-package}
\title{CrimeIncomeDf: Crime and income dataset used to analyze the relationship between income and crime}
\description{
Will update later
}
\keyword{internal}
