% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_joined_df.R
\docType{data}
\name{Inner_joined_df}
\alias{Inner_joined_df}
\title{Inner joined dataframe 1}
\format{
\describe{
}
}
\source{
\url{}
}
\usage{
Inner_joined_df
}
\description{
Inner joined dataframe
}
\keyword{datasets}
